Table을 float type의 columnar vector로 변경하고 DatabaseLoader를 없앰

+++ src/server/table.h
@@ -3,6 +3,10 @@
 
 #include <vector>
 
-using Table = std::vector<std::vector<int>>;
+/* columnar table
+ * Table
+ * [[column 1], [column 2], [column 3], ...]
+ */ 
+using Table = std::vector<std::vector<float>>;
 
 #endif /* EXECUTOR_TRAINER_COMMON_H_ */
